requirements.txt
# Base requirements
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic==1.10.8
python-multipart>=0.0.6
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.2
PyYAML>=6.0
psutil>=5.9.5
apscheduler>=3.10.1
SQLAlchemy>=2.0.12

# Performance optimizations for large files
dask[complete]>=2023.4.1
pyarrow>=12.0.0
numba>=0.58.0
fsspec>=2023.4.0
distributed>=2023.4.1
tqdm>=4.65.0

# Database drivers
psycopg2-binary>=2.9.6

# Excel support
xlsxwriter>=3.1.0
xlrd>=2.0.1

# Memory and system monitoring
memory-profiler>=0.61.0

# Optimized CSV and parquet handling
fastparquet>=2023.4.0
python-snappy>=0.6.1

# Parallel processing
joblib>=1.2.0
uvloop>=0.14.0
httptools>=0.1.1pydantic-settings
